simplify/standardize region setting
If we stay with squishy long term, we really
should either add a different #aws/region
data-reader tag, or expand the current one to work with both common string formats for a
region. Right now it only works with uppercase
and underscores, but the other format is lowercase and dashes. Because region is a
shared setting, it's easiest to have it as the
latter format, and then I have use Java interop
code in here to construct it the other way.